fix: 🐛 EventListeners

+++ thorvg-android/src/main/java/com/lottiefiles/dotlottie/core/widget/DotLottieEventListener.kt
@@ -2,7 +2,7 @@ package com.lottiefiles.dotlottie.core.widget
 
 interface DotLottieEventListener {
     fun onStop() {}
-    fun onFrame(frame: Double) {}
+    fun onFrame(frame: Float) {}
     fun onPause() {}
     fun onPlay() {}
     fun onDestroy() {}
